# frontend/openapi.regulus.json
{
  "components": {
    "schemas": {
      "APIMessage": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "APIMessage",
        "type": "object"
      },
...
      "ConnectorConfigOut": {
        "properties": {
          "config_json": {
            "additionalProperties": true,
            "title": "Config Json",
            "type": "object"
          },
          "connector_type": {
            "enum": [
              "langfuse",
              "prometheus",
              "otel_metrics",
              "posthog",
              "kafka",
              "segment",
              "snowplow",
              "clickhouse",
              "bigquery",
              "snowflake",
              "launchdarkly",
              "envoy"
            ],
            "title": "Connector Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "connector_type",
          "enabled",
          "config_json",
          "created_at",
          "updated_at"
        ],
        "title": "ConnectorConfigOut",
        "type": "object"
      },
      "ConnectorConfigRequest": {
        "properties": {
          "config_json": {
            "additionalProperties": true,
            "title": "Config Json",
            "type": "object"
          }
        },
        "title": "ConnectorConfigRequest",
        "type": "object"
      },
      "ConnectorOutboxOut": {
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_key": {
            "title": "Event Key",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "next_attempt_at": {
            "format": "date-time",
            "title": "Next Attempt At",
            "type": "string"
          },
          "payload_json": {
            "additionalProperties": true,
            "title": "Payload Json",
            "type": "object"
          },
          "processed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processed At"
          },
          "status": {
            "enum": [
              "PENDING",
              "PROCESSING",
              "SENT",
              "FAILED",
              "DEAD_LETTER"
            ],
            "title": "Status",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenant_id",
          "event_type",
          "event_key",
          "payload_json",
          "status",
          "attempts",
          "next_attempt_at",
          "last_error",
          "created_at",
          "processed_at"
        ],
        "title": "ConnectorOutboxOut",
        "type": "object"
      },
      "ConnectorStatusOut": {
        "properties": {
          "connector_type": {
            "enum": [
              "langfuse",
              "prometheus",
              "otel_metrics",
              "posthog",
              "kafka",
              "segment",
              "snowplow",
              "clickhouse",
              "bigquery",
              "snowflake",
              "launchdarkly",
              "envoy"
            ],
            "title": "Connector Type",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "healthy": {
            "title": "Healthy",
            "type": "boolean"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "connector_type",
          "enabled",
          "healthy",
          "message"
        ],
        "title": "ConnectorStatusOut",
        "type": "object"
      },
...
      "EvaluationRunRequest": {
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "confidence_level": {
            "default": 0.95,
            "title": "Confidence Level",
            "type": "number"
          },
          "implementation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implementation Id"
          },
          "mode": {
            "enum": [
              "AB_TEST",
              "SHADOW",
              "PROXY_MODEL"
            ],
            "title": "Mode",
            "type": "string"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          }
        },
        "required": [
          "capability_id",
          "mode",
          "period_start",
          "period_end"
        ],
        "title": "EvaluationRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetryOutboxResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "status": {
            "enum": [
              "PENDING",
              "PROCESSING",
              "SENT",
              "FAILED",
              "DEAD_LETTER"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status"
        ],
        "title": "RetryOutboxResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ValueEstimateOut": {
        "properties": {
          "capability_id": {
            "title": "Capability Id",
            "type": "string"
          },
          "confidence_breakdown": {
            "additionalProperties": true,
            "title": "Confidence Breakdown",
            "type": "object"
          },
          "confidence_gate_passed": {
            "title": "Confidence Gate Passed",
            "type": "boolean"
          },
          "confidence_level": {
            "title": "Confidence Level",
            "type": "number"
          },
          "cost_data_quality": {
            "title": "Cost Data Quality",
            "type": "string"
          },
          "credible_interval_high_usd": {
            "title": "Credible Interval High Usd",
            "type": "number"
          },
          "credible_interval_low_usd": {
            "title": "Credible Interval Low Usd",
            "type": "number"
          },
          "drift_score": {
            "title": "Drift Score",
            "type": "number"
          },
          "drift_state": {
            "enum": [
              "stable",
              "warning",
              "critical"
            ],
            "title": "Drift State",
            "type": "string"
          },
          "estimated_cost_usd": {
            "title": "Estimated Cost Usd",
            "type": "number"
          },
          "estimated_value_usd": {
            "title": "Estimated Value Usd",
            "type": "number"
          },
          "estimation_method_version": {
            "title": "Estimation Method Version",
            "type": "string"
          },
          "implementation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implementation Id"
          },
          "interval_method": {
            "title": "Interval Method",
            "type": "string"
          },
          "method_metadata": {
            "additionalProperties": true,
            "title": "Method Metadata",
            "type": "object"
          },
          "net_margin_usd": {
            "title": "Net Margin Usd",
            "type": "number"
          },
          "period_end": {
            "format": "date-time",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date-time",
            "title": "Period Start",
            "type": "string"
          },
          "relative_interval_width": {
            "title": "Relative Interval Width",
            "type": "number"
          },
          "value_data_quality": {
            "title": "Value Data Quality",
            "type": "string"
          }
        },
        "required": [
          "capability_id",
          "period_start",
          "period_end",
          "estimated_value_usd",
          "estimated_cost_usd",
          "net_margin_usd",
          "credible_interval_low_usd",
          "credible_interval_high_usd",
          "confidence_level",
          "relative_interval_width",
          "confidence_gate_passed",
          "estimation_method_version",
          "cost_data_quality",
          "value_data_quality",
          "drift_score",
          "drift_state",
          "interval_method"
        ],
        "title": "ValueEstimateOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AI Economic Control Plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/connectors": {
      "get": {
        "operationId": "get_connectors_v1_connectors_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectorStatusOut"
                  },
                  "title": "Response Get Connectors V1 Connectors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Connectors",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
    "/v1/connectors/outbox": {
      "get": {
        "operationId": "get_outbox_v1_connectors_outbox_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectorOutboxOut"
                  },
                  "title": "Response Get Outbox V1 Connectors Outbox Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Outbox",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
    "/v1/connectors/outbox/{outbox_id}/retry": {
      "post": {
        "operationId": "retry_outbox_v1_connectors_outbox__outbox_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "outbox_id",
            "required": true,
            "schema": {
              "title": "Outbox Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetryOutboxResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retry Outbox",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
    "/v1/connectors/{connector_type}/configure": {
      "post": {
        "operationId": "configure_v1_connectors__connector_type__configure_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_type",
            "required": true,
            "schema": {
              "title": "Connector Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectorConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Configure",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
    "/v1/connectors/{connector_type}/disable": {
      "post": {
        "operationId": "disable_v1_connectors__connector_type__disable_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_type",
            "required": true,
            "schema": {
              "title": "Connector Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
    "/v1/connectors/{connector_type}/enable": {
      "post": {
        "operationId": "enable_v1_connectors__connector_type__enable_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_type",
            "required": true,
            "schema": {
              "title": "Connector Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectorConfigOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enable",
        "tags": [
          "connectors",
          "metrics"
        ]
      }
    },
...
    "/v1/evaluations/run": {
      "post": {
        "operationId": "run_v1_evaluations_run_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValueEstimateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run",
        "tags": [
          "counterfactual"
        ]
      }
    },
    "/v1/evaluations/run/async": {
      "post": {
        "operationId": "run_async_v1_evaluations_run_async_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIMessage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Async",
        "tags": [
          "counterfactual"
        ]
      }
    },
    "/v1/evaluations/{capability_id}/history": {
      "get": {
        "operationId": "history_v1_evaluations__capability_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ValueEstimateOut"
                  },
                  "title": "Response History V1 Evaluations  Capability Id  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "History",
        "tags": [
          "counterfactual"
        ]
      }
    },
    "/v1/evaluations/{capability_id}/latest": {
      "get": {
        "operationId": "latest_v1_evaluations__capability_id__latest_get",
        "parameters": [
          {
            "in": "path",
            "name": "capability_id",
            "required": true,
            "schema": {
              "title": "Capability Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValueEstimateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Latest",
        "tags": [
          "counterfactual"
        ]
      }
    },
...
  }
}